Use actual generics, move async implementation to utils
